requirements.txt
# --- Core AI & Agent Logic ---
langchain-openai==1.1.10
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-sdk==0.3.9

# --- Database & Vector Store ---
pinecone==8.1.0
pinecone-client==6.0.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
# Official LangChain integration for Pinecone
langchain-pinecone==0.2.13

# --- API & Server ---
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2

# --- Data Processing ---
pypdf==6.7.5
tiktoken==0.12.0

# --- Utils ---
pydantic==2.12.5
httpx==0.28.1